requirements.txt
# GeminiReport 最小限依存関係

# Web Framework
fastapi==0.116.1
uvicorn[standard]==0.35.0
pydantic==2.11.7
flask==3.0.0

# Google Cloud Services
google-cloud-bigquery==3.35.0
google-cloud-storage==2.19.0
google-cloud-pubsub==2.31.0
google-cloud-logging==3.12.1
google-cloud-monitoring==2.27.2
google-cloud-aiplatform==1.95.1

# AI/ML
google-generativeai==0.8.5
# vertexai==1.71.1  # 必要に応じて後で追加

# Google Agent Development Kit
google-adk==1.7.0

# HTTP Client
httpx==0.28.1
aiohttp==3.12.14

# Data Processing
pandas==2.3.1
numpy==2.3.1

# Environment & Config
python-dotenv==1.1.1
pydantic-settings==2.10.1

# Async Support
aiofiles==24.1.0

# Logging
structlog==25.4.0

# Date/Time
python-dateutil==2.9.0.post0
pytz==2025.2

# JSON/YAML
pyyaml==6.0.2  # ADKライブラリが内部で使用

# Markdown Processing
markdown==3.8.2

# 開発用（必要に応じてコメントアウト解除）
# Testing
# pytest>=7.4.0
# pytest-asyncio>=0.21.0
# pytest-cov>=4.1.0

# Development
# black>=23.11.0
# isort>=5.12.0
# flake8>=6.1.0
# mypy>=1.7.0

# Documentation
# mkdocs>=1.5.0
# mkdocs-material>=9.4.0

# Security
# cryptography>=41.0.0

# Image Processing (必要に応じて)
pillow>=10.1.0

# PDF Processing
PyPDF2>=3.0.0
pdfplumber>=0.10.0

# Google Slides API
google-api-python-client>=2.0.0
google-auth-httplib2>=0.1.0
google-auth-oauthlib>=1.0.0

# Data Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.17.0

# Database (BigQuery使用のため不要)
# sqlalchemy>=2.0.0
# alembic>=1.13.0 